+++ x/crosschain/types/errors.go
@@ -41,5 +41,6 @@ var (
 	ErrUnableProcessRefund           = errorsmod.Register(ModuleName, 1148, "unable to process refund")
 	ErrUnableToFindZetaAccounting    = errorsmod.Register(ModuleName, 1149, "unable to find zeta accounting")
 	ErrInsufficientZetaAmount        = errorsmod.Register(ModuleName, 1150, "insufficient zeta amount")
-	ErrMaxTxOutTrackerHashesReached  = errorsmod.Register(ModuleName, 1151, "max tx out tracker hashes reached")
+	ErrInvalidRateLimiterFlags       = errorsmod.Register(ModuleName, 1151, "invalid rate limiter flags")
+	ErrMaxTxOutTrackerHashesReached  = errorsmod.Register(ModuleName, 1152, "max tx out tracker hashes reached")
 )